move renderer to core and clean __init__
