+++ mmhuman3d/core/renderer/torch3d_renderer/lights/__init__.py
@@ -1,0 +1,10 @@
+# yapf: disable
+from .builder import (  # noqa: F401
+    AmbientLights,
+    DirectionalLights,
+    PointLights,
+    build_lights,
+)
+from .lights import MMLights  # noqa: F401
+
+# yapf: enable
