+++ mmhuman3d/core/renderer/torch3d_renderer/segmentation_renderer.py
@@ -4,27 +4,31 @@ import torch
 from pytorch3d.structures import Meshes
 
 from mmhuman3d.core.cameras import MMCamerasBase
+from mmhuman3d.utils import get_different_colors
 from .base_renderer import BaseRenderer
-from .builder import RENDERER
 from .utils import normalize
 
 
-@RENDERER.register_module(name=[
-    'silhouette', 'silhouette_renderer', 'Silhouette', 'SilhouetteRenderer'
-])
-class SilhouetteRenderer(BaseRenderer):
-    """Silhouette renderer."""
-    shader_type = 'SilhouetteShader'
+class SegmentationRenderer(BaseRenderer):
+    """Render segmentation map into a segmentation index tensor."""
+    shader_type = 'SegmentationShader'
 
-    def __init__(
-        self,
-        resolution: Tuple[int, int] = None,
-        device: Union[torch.device, str] = 'cpu',
-        output_path: Optional[str] = None,
-        out_img_format: str = '%06d.png',
-        **kwargs,
-    ) -> None:
-        """SilhouetteRenderer for neural rendering and visualization.
+    def __init__(self,
+                 resolution: Tuple[int, int] = None,
+                 device: Union[torch.device, str] = 'cpu',
+                 output_path: Optional[str] = None,
+                 out_img_format: str = '%06d.png',
+                 num_class: int = 1,
+                 **kwargs) -> None:
+        """Render vertex-color mesh into a segmentation map of a (B, H, W)
+        tensor. For visualization, the output rgba image will be (B, H, W, 4),
+        and the color palette comes from `get_different_colors`. The
+        segmentation map is a tensor each pixel saves the classification index.
+        Please make sure you have allocate each pixel a correct classification
+        index by defining a textures of vertex color.
+
+        [CrossEntropyLoss](https://pytorch.org/docs/stable/generated/torch.nn.
+        CrossEntropyLoss.html)
 
         Args:
             resolution (Iterable[int]):
@@ -38,6 +42,8 @@ class SilhouetteRenderer(BaseRenderer):
             out_img_format (str, optional): The image format string for
                 saving the images.
                 Defaults to '%06d.png'.
+            num_class (int, optional): number of segmentation parts.
+                Defaults to 1.
 
         Returns:
             None
@@ -46,23 +52,23 @@ class SilhouetteRenderer(BaseRenderer):
             resolution=resolution,
             device=device,
             output_path=output_path,
+            obj_path=None,
             out_img_format=out_img_format,
             **kwargs)
+        self.num_class = num_class
 
     def forward(self,
-                meshes: Optional[Meshes] = None,
+                meshes: Meshes,
                 cameras: Optional[MMCamerasBase] = None,
-                images: Optional[torch.Tensor] = None,
-                indexes: Iterable[str] = None,
+                indexes: Optional[Iterable[int]] = None,
                 backgrounds: Optional[torch.Tensor] = None,
                 **kwargs):
-        """Render silhouette map.
+        """Render segmentation map.
 
         Args:
-            meshes (Optional[Meshes], optional): meshes to be rendered.
+            meshes (Meshes): meshes to be rendered.
                 Require the textures type is `TexturesClosest`.
                 The color indicates the class index of the triangle.
-                Defaults to None.
             cameras (Optional[MMCamerasBase], optional): cameras for render.
                 Defaults to None.
             indexes (Optional[Iterable[int]], optional): indexes for images.
@@ -73,21 +79,27 @@ class SilhouetteRenderer(BaseRenderer):
         Returns:
             Union[torch.Tensor, None]: return tensor or None.
         """
+
         meshes = meshes.to(self.device)
         self._update_resolution(cameras, **kwargs)
         fragments = self.rasterizer(meshes_world=meshes, cameras=cameras)
-        silhouette_map = self.shader(
+        segmentation_map = self.shader(
             fragments=fragments, meshes=meshes, cameras=cameras)
 
         if self.output_path is not None:
-            rgba = self.tensor2rgba(silhouette_map)
-            self._write_images(rgba, backgrounds, indexes)
+            rgba = self.tensor2rgba(segmentation_map)
+            if self.output_path is not None:
+                self._write_images(rgba, backgrounds, indexes)
 
-        return silhouette_map
+        return segmentation_map
 
     def tensor2rgba(self, tensor: torch.Tensor):
-        silhouette = tensor[..., 3:]
-        rgbs = silhouette.repeat(1, 1, 1, 3)
-        valid_masks = (silhouette > 0) * 1.0
-        rgbs = normalize(rgbs, out_value_range=(0, 1))
-        return torch.cat([rgbs, valid_masks], -1)
+        valid_masks = (tensor[..., :] > 0) * 1.0
+        color = torch.Tensor(get_different_colors(self.num_class))
+        color = torch.cat([torch.zeros(1, 3), color]).to(self.device)
+        B, H, W, _ = tensor.shape
+        rgbs = color[tensor.view(-1)].view(B, H, W, 3) * valid_masks
+        rgbs = normalize(
+            rgbs.float(), origin_value_range=(0, 255), out_value_range=(0, 1))
+        rgba = torch.cat([rgbs, valid_masks], -1)
+        return rgba
