+++ mmhuman3d/core/renderer/torch3d_renderer/textures/__init__.py
@@ -1,0 +1,10 @@
+# yapf: disable
+from .builder import (  # noqa:F401
+    TexturesAtlas,
+    TexturesNearest,
+    TexturesUV,
+    TexturesVertex,
+    build_textures,
+)
+
+# yapf: enable
