+++ mmhuman3d/core/visualization/__init__.py
@@ -1,24 +1,2 @@
-from .renderer.torch3d_renderer import render_runner
-from .renderer.torch3d_renderer.builder import (
-    build_lights,
-    build_renderer,
-    build_shader,
-    build_textures,
-)
-from .visualize_keypoints2d import visualize_kp2d
-from .visualize_keypoints3d import visualize_kp3d
-from .visualize_smpl import (
-    render_smpl,
-    visualize_smpl_calibration,
-    visualize_smpl_hmr,
-    visualize_smpl_pose,
-    visualize_smpl_vibe,
-    visualize_T_pose,
-)
-
-__all__ = [
-    'visualize_kp2d', 'visualize_kp3d', 'visualize_smpl_pose',
-    'visualize_T_pose', 'render_smpl', 'visualize_smpl_vibe',
-    'visualize_smpl_calibration', 'visualize_smpl_hmr', 'render_runner',
-    'build_lights', 'build_renderer', 'build_shader', 'build_textures'
-]
+from .visualize_keypoints2d import visualize_kp2d  # noqa:F401
+from .visualize_keypoints3d import visualize_kp3d  # noqa:F401
