--- mmhuman3d/core/visualization/renderer/__init__.py
@@ -1,34 +1,0 @@
-from .matplotlib3d_renderer import Axes3dJointsRenderer
-from .torch3d_renderer import render_runner
-from .torch3d_renderer.base_renderer import BaseRenderer
-from .torch3d_renderer.builder import (
-    build_lights,
-    build_renderer,
-    build_shader,
-    build_textures,
-)
-from .torch3d_renderer.depth_renderer import DepthRenderer
-from .torch3d_renderer.mesh_renderer import MeshRenderer
-from .torch3d_renderer.normal_renderer import NormalRenderer
-from .torch3d_renderer.pointcloud_renderer import PointCloudRenderer
-from .torch3d_renderer.segmentation_renderer import SegmentationRenderer
-from .torch3d_renderer.shader import (
-    DepthShader,
-    NoLightShader,
-    NormalShader,
-    SegmentationShader,
-)
-from .torch3d_renderer.silhouette_renderer import SilhouetteRenderer
-from .torch3d_renderer.smpl_renderer import SMPLRenderer
-from .torch3d_renderer.textures import TexturesNearest
-from .torch3d_renderer.uv_renderer import UVRenderer
-from .vedo_render import VedoRenderer
-
-__all__ = [
-    'NoLightShader', 'BaseRenderer', 'TexturesNearest', 'SMPLRenderer',
-    'SilhouetteRenderer', 'Axes3dJointsRenderer', 'VedoRenderer',
-    'MeshRenderer', 'DepthRenderer', 'NormalRenderer', 'SegmentationRenderer',
-    'PointCloudRenderer', 'UVRenderer', 'build_renderer', 'build_shader',
-    'build_textures', 'build_lights', 'DepthShader', 'SegmentationShader',
-    'NormalShader', 'NoLightShader', 'render_runner'
-]
