+++ tests/test_vis/test_uv.py
@@ -1,7 +1,7 @@
 import torch
 from pytorch3d.structures import Meshes
 
-from mmhuman3d.core.visualization.renderer import UVRenderer
+from mmhuman3d.core.renderer.torch3d_renderer.builder import UVRenderer
 from mmhuman3d.models.body_models.builder import build_body_model
 
 if torch.cuda.is_available():
